requirements.txt
amqp==5.3.1
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.8
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.2.1
Django==5.2
django-celery-beat==2.8.1
django-filter==25.1
django-timezone-field==7.1
djangorestframework-simplejwt==5.5.0
djangorestframework==3.16.0
drf-yasg==1.21.10
forex-python==1.8
idna==3.10
inflection==0.5.1
kombu==5.5.3
packaging==25.0
psycopg2-binary==2.9.10
pyjwt==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
Pillow==11.2.1